1. 数据导出接口完成

+++ sky-server/src/main/java/com/sky/mapper/UserMapper.java
@@ -14,5 +14,5 @@ public interface UserMapper {
     //查询指定日期时间段内的所有用户人数
     public long userAllStatMapper(LocalDateTime begin,LocalDateTime end);
     //统计指定日期时间段内新增加的用户数量
-    public long userNewAddStatMapper(LocalDateTime begin,LocalDateTime end);
+    public Integer userNewAddStatMapper(LocalDateTime begin,LocalDateTime end);
 }
